requirements-runpod.txt
# LiveKit Agent Framework
livekit>=1.0.0
livekit-agents>=1.3.0

# LiveKit Plugins
livekit-plugins-deepgram>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-silero>=1.0.0

# LLM
anthropic>=0.40.0

# STT - Opzione 2: Whisper locale su GPU
faster-whisper>=1.0.0
ctranslate2>=4.0.0
numpy

# STT - Opzione 3: Whisper API (HTTP)
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
